GOOGLE-48
Simplified code and replace use of homemade escapeJson() function on each data item with standard JSON.stringify() method on the entire data object.

The same should be done with code in the addPermissionsToFile() function, too.  If that function were cleaned up, there would no longer be a need for escapeJson(), which is implemented in utils.js.