GOOGLE-80
With some small alteration, the code used to display the contents of a shared folder to students is also used to display a tree of folders to the instructor, from which they can choose a folder to share.  After this change, some of the old special-purpose folder list code should be removed, since it's now unused and obsolete.