GOOGLE-155
getDriveFile(): Surprisingly, this method has been using the error callback's arguments *in the wrong order* for a long time.  I've fixed this and also changed the code to use the "promise" model in an attempt to improve error handling.  It didn't help, but it didn't hurt.  The promise model is recommended for new jQuery code, so I'll keep this code.

_getGoogleDriveUrl(): Simplified and clarified by using meaningful, descriptive variable names.

Also, JSLint recommended changes and simplified some other code.