GOOGLE-158
New files:

about.html - Contents of the new modal About dialog box.

index.html - Page shown when the application is called without any additional path or parameters.  Used to let administrators see that the application is installed and running.

build.txt - A file of information about the current build of the application.  Referenced by about.html and index.html.  This should be filled in by the developer building the application, preferably by an automated build script.